test: split test of NumberBallSetFactory
NumberBallSetFactory의 검증 로직을 Validator로 분리함에 따라
검증 로직을 validator로 이동
